Rename game-modes/ to modes/
It was bothering me
