+++ lib/game/manager.js
@@ -9,7 +9,7 @@ const events = require('events')
 const debug = require('debug')('colonialwars:manager')
 
 const MapConfig = require('./map-config')
-const games = require('./game-modes')
+const games = require('./modes')
 const { COMMUNICATIONS: communications } = require('../constants')
 
 /**
@@ -30,8 +30,8 @@ const { COMMUNICATIONS: communications } = require('../constants')
  * @prop {number} currentPlayers
  */
 /**
- * @typedef {import('./game-modes/base-game').WSConnInstance} WSConnInstance
- * @typedef {import('./game-modes/base-game').PlayerMeta} PlayerMeta
+ * @typedef {import('./modes/base-game').WSConnInstance} WSConnInstance
+ * @typedef {import('./modes/base-game').PlayerMeta} PlayerMeta
  * @typedef {import('./player').PlayerInput} PlayerInput
  *
  * @typedef {Record<'id'|'name'|'mode'|'description', string>} GameInfo
@@ -89,7 +89,7 @@ class Manager extends events.EventEmitter {
     /**
      * This is a Map containing all of the games running
      * on the server.
-     * @type {Map<string, import('./game-modes/base-game')>}
+     * @type {Map<string, import('./modes/base-game')>}
      * @private
      */
     this._games = new Map()
@@ -155,7 +155,7 @@ class Manager extends events.EventEmitter {
    * Adds a new client to a game.
    * @param {string} gameID The game's ID.
    * @param {InstanceType<import('../cwdtp/conn')>} conn The WSConn object associated with the client.
-   * @param {import('./game-modes/base-game').PlayerMeta} playerMeta Player information.
+   * @param {import('./modes/base-game').PlayerMeta} playerMeta Player information.
    */
   addClientTo (gameID, conn, playerMeta) {
     if (this._games.has(gameID)) {
