+++ lib/game/modes/base-game.js
