+++ lib/game/modes/index.js
