+++ lib/game/modes/team-game.js
